updated try catch in controller admin page

+++ resources/views/admin/user/index.blade.php
@@ -33,10 +33,9 @@
                     </td>
                     <td>
                         <a href="{{ route('admin.user.edit', $user->id) }}" class="btn btn-primary btn-sm mego-button-edit">Edit</a>
-                        {!! Form::open(['route' => ['admin.user.destroy', $user->id], 'onsubmit' => "return confirm('Are
-                            you sure?')", 'method' => 'DELETE']) !!}
+                         {!! Form::open(['route' => ['admin.user.destroy', $user->id], 'onsubmit' => "return confirm('Are you sure?')", 'method' => 'DELETE']) !!}
                             <button type="submit" class="btn btn-danger btn-sm">Delete</button>
-                        {!! Form::close() !!}
+                         {!! Form::close() !!}
                     </td>
                 </tr>
                 @endforeach
